Move fixtures up one level
